# ToTitleCase Perf improvements