--- a/src/Benchmarks/TransformersBenchmarks.cs
+++ b/src/Benchmarks/TransformersBenchmarks.cs
@@ -5,12 +5,21 @@ public class TransformersBenchmarks
     const int RAND_SEED = 17432;
 
     static readonly char[] _alphabet =
-        Enumerable.Repeat((int)' ', 12)
+        Enumerable
+            .Repeat((int) ' ', 12)
             .Concat(Enumerable.Range('a', 'z' - 'a'))
             .Concat(Enumerable.Range('A', 'Z' - 'A'))
             .Concat(Enumerable.Range('0', '9' - '0'))
-            .Concat(new int[] { '.', ',', '(', ')', '!', '$' })
-            .Select(x => (char)x)
+            .Concat(new int[]
+            {
+                '.',
+                ',',
+                '(',
+                ')',
+                '!',
+                '$'
+            })
+            .Select(x => (char) x)
             .ToArray();
 
     Random random = new(RAND_SEED);
@@ -27,24 +36,25 @@ public class TransformersBenchmarks
         {
             chars[i] = _alphabet[random.Next(0, _alphabet.Length)];
         }
+
         input = new(chars);
     }
-    //
-    // [Benchmark]
-    // public string AllTransforms() =>
-    //     input.Transform(To.LowerCase, To.UpperCase, To.SentenceCase, To.TitleCase);
-    //
-    // [Benchmark]
-    // public string LowerCase() =>
-    //     input.Transform(To.LowerCase);
-    //
-    // [Benchmark]
-    // public string UpperCase() =>
-    //     input.Transform(To.UpperCase);
-    //
-    // [Benchmark]
-    // public string SentenceCase() =>
-    //     input.Transform(To.SentenceCase);
+
+    [Benchmark]
+    public string AllTransforms() =>
+        input.Transform(To.LowerCase, To.UpperCase, To.SentenceCase, To.TitleCase);
+
+    [Benchmark]
+    public string LowerCase() =>
+        input.Transform(To.LowerCase);
+
+    [Benchmark]
+    public string UpperCase() =>
+        input.Transform(To.UpperCase);
+
+    [Benchmark]
+    public string SentenceCase() =>
+        input.Transform(To.SentenceCase);
 
     [Benchmark]
     public string TitleCase() =>