--- a/src/Humanizer/Transformer/ToTitleCase.cs
+++ b/src/Humanizer/Transformer/ToTitleCase.cs
@@ -1,4 +1,6 @@
-﻿namespace Humanizer
+﻿using System.Collections.Frozen;
+
+namespace Humanizer
 {
     class ToTitleCase : ICulturedStringTransformer
     {
@@ -27,19 +29,24 @@
         static bool AllCapitals(string input) =>
             input.All(char.IsUpper);
 
-        static string ReplaceWithTitleCase(Match word, string source, CultureInfo culture, bool firstWord)
+        static FrozenSet<string> lookups;
+
+        static ToTitleCase()
         {
             var articles = new List<string> { "a", "an", "the" };
             var conjunctions = new List<string> { "and", "as", "but", "if", "nor", "or", "so", "yet" };
             var prepositions = new List<string> { "as", "at", "by", "for", "in", "of", "off", "on", "to", "up", "via" };
 
+            lookups = articles.Concat(conjunctions).Concat(prepositions).ToFrozenSet();
+        }
+
+        static string ReplaceWithTitleCase(Match word, string source, CultureInfo culture, bool firstWord)
+        {
             var wordToConvert = word.Value;
             string replacement;
 
             if (firstWord ||
-                (!articles.Contains(wordToConvert) &&
-                !conjunctions.Contains(wordToConvert) &&
-                !prepositions.Contains(wordToConvert)))
+                !lookups.Contains(wordToConvert))
             {
                 replacement = culture.TextInfo.ToUpper(wordToConvert[0]) + culture.TextInfo.ToLower(wordToConvert.Remove(0, 1));