v2

| Method    | StringLen | Mean      | Error     | StdDev    | Allocated |
|---------- |---------- |----------:|----------:|----------:|----------:|
| TitleCase | 10        |  2.283 us | 0.0055 us | 0.0052 us |   1.57 KB |
| TitleCase | 100       | 34.372 us | 0.0438 us | 0.0389 us |  24.85 KB |


v3

| Method    | StringLen | Mean      | Error     | StdDev    | Allocated |
|---------- |---------- |----------:|----------:|----------:|----------:|
| TitleCase | 10        |  1.008 us | 0.0030 us | 0.0028 us |     920 B |
| TitleCase | 100       | 10.872 us | 0.0130 us | 0.0115 us |    8720 B |